net: Use skbufhead with raw lock
Use the rps lock as rawlock so we can keep irq-off regions. It looks low
latency. However we can't kfree() from this context therefore we defer this
to the softirq and use the tofree_queue list for it (similar to process_queue).

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
(cherry picked from commit 16d85330ef503daccbf8c622c9141acd40262d86)